backend/app/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
python-dotenv
anthropic
cors
pdfplumber
python-docx
google-generativeai
PyPDF2
groq
requests
openai